+++ ChangeLog.md
@@ -5,6 +5,7 @@
 * Drain de force de Dominia (Invincible)
 
 ### Autres améliorations
+* Ajout d'une options pour des DM explosifs.
 * Ajout de l'option --bonusContreArmure
 * Changement des macros Monter et Descendre en icones.
 * Changement de fonctionnement des auras, qui réagissent maintenant au déplacement